requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
slowapi
python-dotenv